refactor: extract from directory
