+++ src/main/kotlin/ecommerce/dto/PaymentRequest.kt
@@ -1,4 +1,4 @@
-package ecommerce.dto.stripe
+package ecommerce.dto
 
 class PaymentRequest(
     val amount: Int,
