+++ src/test/kotlin/ecommerce/stripe/StripeClientTest.kt
@@ -1,6 +1,6 @@
 package ecommerce.stripe
 
-import ecommerce.dto.stripe.PaymentRequest
+import ecommerce.dto.PaymentRequest
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
